Feature/purchases

+++ src/validators/productPrices.js
@@ -7,7 +7,8 @@ const validateGetRecord = [
   check('id')
     .exists().withMessage(PRODUCT_PRICES_VALIDATORS.ID_NOT_EXISTS).bail()
     .notEmpty().withMessage(PRODUCT_PRICES_VALIDATORS.ID_IS_EMPTY).bail()
-    .isInt().withMessage(PRODUCT_PRICES_VALIDATORS.ID_INVALID).bail(),
+    .isInt().withMessage(PRODUCT_PRICES_VALIDATORS.ID_INVALID).bail()
+    .toInt(),
   (req, res, next) => {
     return validateResults(req, res, next);
   }
@@ -17,7 +18,8 @@ const validateGetByProduct = [
   check('product_id')
     .exists().withMessage(PRODUCT_PRICES_VALIDATORS.PRODUCT_ID_NOT_EXISTS).bail()
     .notEmpty().withMessage(PRODUCT_PRICES_VALIDATORS.PRODUCT_ID_IS_EMPTY).bail()
-    .isInt().withMessage(PRODUCT_PRICES_VALIDATORS.PRODUCT_ID_INVALID).bail(),
+    .isInt().withMessage(PRODUCT_PRICES_VALIDATORS.PRODUCT_ID_INVALID).bail()
+    .toInt(),
   (req, res, next) => {
     return validateResults(req, res, next);
   }
@@ -27,7 +29,8 @@ const validateGetByPriceList = [
   check('price_list_id')
     .exists().withMessage(PRODUCT_PRICES_VALIDATORS.PRICE_LIST_ID_NOT_EXISTS).bail()
     .notEmpty().withMessage(PRODUCT_PRICES_VALIDATORS.PRICE_LIST_ID_IS_EMPTY).bail()
-    .isInt().withMessage(PRODUCT_PRICES_VALIDATORS.PRICE_LIST_ID_INVALID).bail(),
+    .isInt().withMessage(PRODUCT_PRICES_VALIDATORS.PRICE_LIST_ID_INVALID).bail()
+    .toInt(),
   (req, res, next) => {
     return validateResults(req, res, next);
   }
@@ -37,11 +40,13 @@ const valiAddRecord = [
   check('product_id')
     .exists().withMessage(PRODUCT_PRICES_VALIDATORS.PRODUCT_ID_NOT_EXISTS).bail()
     .notEmpty().withMessage(PRODUCT_PRICES_VALIDATORS.PRODUCT_ID_IS_EMPTY).bail()
-    .isInt().withMessage(PRODUCT_PRICES_VALIDATORS.PRODUCT_ID_INVALID).bail(),
+    .isInt().withMessage(PRODUCT_PRICES_VALIDATORS.PRODUCT_ID_INVALID).bail()
+    .toInt(),
   check('price_list_id')
     .exists().withMessage(PRODUCT_PRICES_VALIDATORS.PRICE_LIST_ID_NOT_EXISTS).bail()
     .notEmpty().withMessage(PRODUCT_PRICES_VALIDATORS.PRICE_LIST_ID_IS_EMPTY).bail()
-    .isInt().withMessage(PRODUCT_PRICES_VALIDATORS.PRICE_LIST_ID_INVALID).bail(),
+    .isInt().withMessage(PRODUCT_PRICES_VALIDATORS.PRICE_LIST_ID_INVALID).bail()
+    .toInt(),
   check('price')
     .exists().withMessage(PRODUCT_PRICES_VALIDATORS.PRICE_NOT_EXISTS).bail()
     .notEmpty().withMessage(PRODUCT_PRICES_VALIDATORS.PRICE_IS_EMPTY).bail()
@@ -58,13 +63,16 @@ const valiUpdateRecord = [
   check('id')
     .exists().withMessage(PRODUCT_PRICES_VALIDATORS.ID_NOT_EXISTS).bail()
     .notEmpty().withMessage(PRODUCT_PRICES_VALIDATORS.ID_IS_EMPTY).bail()
-    .isInt().withMessage(PRODUCT_PRICES_VALIDATORS.ID_INVALID).bail(),
+    .isInt().withMessage(PRODUCT_PRICES_VALIDATORS.ID_INVALID).bail()
+    .toInt(),
   check('product_id')
     .optional()
-    .isInt().withMessage(PRODUCT_PRICES_VALIDATORS.PRODUCT_ID_INVALID).bail(),
+    .isInt().withMessage(PRODUCT_PRICES_VALIDATORS.PRODUCT_ID_INVALID).bail()
+    .toInt(),
   check('price_list_id')
     .optional()
-    .isInt().withMessage(PRODUCT_PRICES_VALIDATORS.PRICE_LIST_ID_INVALID).bail(),
+    .isInt().withMessage(PRODUCT_PRICES_VALIDATORS.PRICE_LIST_ID_INVALID).bail()
+    .toInt(),
   check('price')
     .optional()
     .isDecimal().withMessage(PRODUCT_PRICES_VALIDATORS.PRICE_INVALID).bail(),
